+++ src/Stormancer.Plugins/Friends/Stormancer.Server.Plugins.Friends/Changelog.rst
@@ -17,7 +17,9 @@ Changed
 *******
 - Updated dependency to Users to support new abstraction lib.
 - Fixes S2S service locator 
-
+Added
+*****
+- Added block list expiration to block API.
 
 5.0.0.14
 ----------